refactored: Sorted Files
[]Removed Test and Example files
[]Relocated Existing Activity files into new auth directory
