+++ app/src/main/java/com/example/mad_finals_wurmple/auth/SignupActivity.kt
@@ -1,12 +1,13 @@
-package com.example.mad_finals_wurmple
+package com.example.mad_finals_wurmple.auth
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.example.mad_finals_wurmple.R
+import com.example.mad_finals_wurmple.menuActivity
 import com.google.firebase.auth.FirebaseAuth
 
 class signupActivity : AppCompatActivity() {
